PR review changes: Reverted netwrapper to platform agnostic package
